+++ comments.php
@@ -77,3 +77,4 @@ if ( post_password_required() ) {
 	?>
 
 </div><!-- .comments-area -->
+