+++ page.php
@@ -35,6 +35,5 @@ get_header(); ?>
 	<?php get_sidebar( 'content-bottom' ); ?>
 
 </div><!-- .content-area -->
-
 <?php get_sidebar(); ?>
 <?php get_footer(); ?>