+++ searchform.php
@@ -9,7 +9,7 @@
 ?>
 
 <form role="search" method="get" class="search-form" action="<?php echo esc_url( home_url( '/' ) ); ?>">
-	<?php /* C. Seguinot add CSRF protection to search fomr TODO test nonce!*/ wp_nonce_field('search','_wpnonce'); ?>
+	<?php wp_nonce_field('search','_wpnonce'); ?>
 	<label>
 		<span class="screen-reader-text"><?php echo _x( 'Search for:', 'label', 'cnrswebkit' ); ?></span>
 		<input type="search" class="search-field" placeholder="<?php /* translators: this is a Search box placeholder (Search...) */