+++ templates/template agenda.php
@@ -1,15 +1,22 @@
 <?php
 /**
- * The template for displaying list of Événements
+ * Template Name: CNRS WebKit list of events
+ * Template Post Type: post, page
+ *
+ * The template for displaying a list of events
  *
  * @package Atos
  * @subpackage CNRS_Web_Kit
  * @since CNRS Web Kit 1.0
  * 
- * Template Name: Events list
  */
+
+// Translators: Template Name translation.
+__('CNRS WebKit list of events', 'cnrswebkit');
+
 get_header();
-//require_once( get_template_directory() . '/inc/ajax.php' ); 
+// TODO next line commented in V0.3! Is ajax useful?? 
+// require_once( get_template_directory() . '/inc/ajax.php' );  
 ?>
 <div id="primary" class="content-area">
     <main id="main" class="site-main" role="main">
@@ -26,19 +33,15 @@ get_header();
                 $_SESSION['date_month'] = false;
                 $evenement_data = new CnrswebkitPageItemsList('evenement');
                 echo $evenement_data->get_html_filters();
-                $evenement_data->get_html_item_list();
-                echo '<div id="evenement_ajax_container"></div>';
-                ?>
-                <div class="moreEvents"><a page="1" target="#evenement_ajax_container"><?php _e('Display more events','cnrswebkit')?></a></div>
-                <script>
-	                var hideLoadMore = false; // C. Seguinot this was not declared on this page! 
-	                (function ($) {
-                        if (hideLoadMore) {
-                            $('.moreEvents').hide();
-                        }
-                    })(jQuery);
-                </script>
-                <?php
+                if ($evenement_data->has_items() ) {
+                    echo $evenement_data->get_html_item_list();
+                    echo '<div id="evenement_ajax_container"></div>';
+                    ?>
+                    <div class="moreEvents"><a page="1" target="#evenement_ajax_container">Afficher plus d'évènements</a></div>
+                    <?php
+                } else {
+                    echo '<br/><p>'. __('There is currently no event in the present list', 'cnrswebkit') . '</p>';
+                }
                 display_bottom_partenaires();
                 ?>
             </div><!-- .entry-content -->