+++ templates/template emploi.php
@@ -1,15 +1,22 @@
 <?php
 /**
- * The template for displaying list of Emplois
+ * Template Name: CNRS WebKit list of job offer
+ * Template Post Type: post, page
+ *
+ * The template for displaying a list of job offer
  *
  * @package Atos
  * @subpackage CNRS_Web_Kit
  * @since CNRS Web Kit 1.0
  * 
- * Template Name: Recruitments list
  */
+
+// Translators: Template Name translation.
+__('CNRS WebKit list of job offer', 'cnrswebkit');
+
 get_header();
-//require_once( get_template_directory() . '/inc/ajax.php' ); 
+// TODO next line commented in V0.3! Is ajax useful?? 
+// require_once( get_template_directory() . '/inc/ajax.php' );  
 ?>
 <div id="primary" class="content-area">
     <main id="main" class="site-main" role="main">
@@ -27,8 +34,12 @@ get_header();
                 $actualites_data = new CnrswebkitPageItemsList('emploi');
                 echo $actualites_data->get_html_filters();
 				echo $actualites_data->get_pagination();
-                echo $actualites_data->get_html_item_list();
-                echo $actualites_data->get_pagination();
+				if ($actualites_data->has_items() ) {
+				    echo $actualites_data->get_html_item_list();
+				} else {
+				    echo '<br/><p>'. __('There is currently no recruitment offer published', 'cnrswebkit') . '</p>';
+				}
+				echo $actualites_data->get_pagination();
                 display_bottom_partenaires();
                 ?>  
             </div><!-- .entry-content -->